Reorganize platform wizard files
