+++ src/renderer/src/views/x/components/XProgressErrorsOccuredComponent.vue
