+++ src/renderer/src/views/x/wizard/XWizardBuildOptionsPage.vue
@@ -1,11 +1,11 @@
 <script setup lang="ts">
 import { ref, computed, onMounted } from "vue";
-import { XViewModel, State } from "../../view_models/XViewModel";
-import { setJobsType, getBreadcrumbIcon } from "../../util";
-import type { StandardWizardPageProps } from "../../types/WizardPage";
-import { useWizardPage } from "../../composables/useWizardPage";
-import BaseWizardPage from "../shared_components/wizard/BaseWizardPage.vue";
-import XLastImportOrBuildComponent from "./XLastImportOrBuildComponent.vue";
+import { XViewModel, State } from "../../../view_models/XViewModel";
+import { setJobsType, getBreadcrumbIcon } from "../../../util";
+import type { StandardWizardPageProps } from "../../../types/WizardPage";
+import { useWizardPage } from "../../../composables/useWizardPage";
+import BaseWizardPage from "../../shared_components/wizard/BaseWizardPage.vue";
+import XLastImportOrBuildComponent from "../components/XLastImportOrBuildComponent.vue";
 
 // Props
 interface Props extends StandardWizardPageProps {
